bin paths from env

+++ bonus/clean_exit_bonus.c
@@ -6,7 +6,7 @@
 /*   By: epilar <epilar@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/19 12:53:55 by epilar            #+#    #+#             */
-/*   Updated: 2022/05/19 12:54:19 by epilar           ###   ########.fr       */
+/*   Updated: 2022/05/19 13:20:08 by epilar           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,13 +29,24 @@ void	print_error(char *msg)
 
 void	clean_exit(t_pipex *pipex, char *msg)
 {
-	print_error(msg);
-	if (pipex->filelst & INFD)
-		close(pipex->infile);
-	if (pipex->filelst & HDFD)
-		unlink(HEREDOC_FILE);
-	if (pipex->filelst & OUTFD)
-		close(pipex->outfile);
+	int	i;
 
+	print_error(msg);
+	if (pipex)
+	{
+		if (pipex->filelst & INFD)
+			close(pipex->infile);
+		if (pipex->filelst & HDFD)
+			unlink(HEREDOC_FILE);
+		if (pipex->filelst & OUTFD)
+			close(pipex->outfile);
+		if (pipex->cmd_paths)
+		{
+			i = 0;
+			while (pipex->cmd_paths[i])
+				free(pipex->cmd_paths[i++]);
+			free(pipex->cmd_paths);
+		}
+	}
 	exit(1);
 }
